+++ lib/simple_jwt_manager.dart
@@ -9,3 +9,5 @@ export 'src/server_response/server_response.dart';
 export 'src/server_response/util_server_response.dart';
 export 'src/ropc_client_stub.dart'
     if (dart.library.js_interop) 'src/ropc_client_web.dart';
+export 'src/tools/error_reporter_stub.dart'
+    if (dart.library.js_interop) 'src/tools/error_reporter_web.dart';