+++ lib/src/tools/error_reporter_stub.dart
@@ -1,0 +1,2 @@
+export 'error_reporter.dart';
+export 'error_reporter_for_native.dart';